+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The main additions are the new assertions for
 ### Added
 
 - Add `True`, `False`, `NoError` assertion functions.
+- Add `Panics`, `NoPanic` assertion functions.
+- Add `Eventually`, `EventuallyChan` assertion functions.
 - Add `Ordered` function which provides following assertions,
   in addition to `Comparable`, via `FluentOrdered` type:
   - `Lesser`